Revert "Revert "feat: get offline login started""
This reverts commit cb8d7803db709cd1fb57f5f1eed874ceb28ebfa7.
