+++ modules/state-management/reducers/root-reducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
 import login from './login';
-import dashboardControls from './dashboardControls';
+import dashboardControls from './dashboard-controls';
 
 export default combineReducers({
   login,
